avances en centro control

+++ src/com/dreamer8/yosimce/shared/dto/CentroOperacionDTO.java
@@ -22,8 +22,8 @@ public class CentroOperacionDTO implements Serializable {
 	private Integer idComuna;
 	private Integer idZona;
 	private String nombre;
-	private Long longitud;
-	private Long latitud;
+	private Double longitud;
+	private Double latitud;
 	private String nombreJefeCentro;
 	private String telefonoJefeCentro;
 	private Integer enCentro;
@@ -100,19 +100,19 @@ public class CentroOperacionDTO implements Serializable {
 		this.telefonoJefeCentro = telefonoJefeCentro;
 	}
 
-	public Long getLongitud() {
+	public Double getLongitud() {
 		return longitud;
 	}
 
-	public void setLongitud(Long longitud) {
+	public void setLongitud(Double longitud) {
 		this.longitud = longitud;
 	}
 
-	public Long getLatitud() {
+	public Double getLatitud() {
 		return latitud;
 	}
 
-	public void setLatitud(Long latitud) {
+	public void setLatitud(Double latitud) {
 		this.latitud = latitud;
 	}
